pyproject.toml
[project]
name = "canvas-archive"
version = "0.1.0"
description = "Pulls Canvas LMS course material (assignment .md per group, optional handouts) into per-course private GitHub repos."
requires-python = ">=3.12"
dependencies = [
    "canvasapi>=3.6.0",
    "markdownify>=1.2.2",
    "pypdf>=6.10.2",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
]

[project.scripts]
canvas-archive = "canvas_archive.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/canvas_archive"]
